# [SR-4698] Optimize getting block locations with a batch api