--- a/fe/fe-core/src/main/java/com/starrocks/external/hive/HiveMetaClient.java
+++ b/fe/fe-core/src/main/java/com/starrocks/external/hive/HiveMetaClient.java
@@ -445,7 +445,7 @@ public class HiveMetaClient {
         URI uri = new URI(dirPath);
         FileSystem fileSystem = getFileSystem(uri);
         // fileSystem.listLocatedStatus is a batch api include fileSystem.listStatus
-        // and ileSystem.getFileBlockLocations two api's information.
+        // and fileSystem.getFileBlockLocations two api's information.
         // the performance is well in get many file located information scenario.
         RemoteIterator<LocatedFileStatus> blockIterator = fileSystem.listLocatedStatus(new Path(uri.getPath()));
         List<HdfsFileDesc> fileDescs = Lists.newArrayList();