listLocatedStatus is an api to list all statuses and block locations of the files in the given path in one operation. 
The performance is better than getting status and block location one by one.
Performance test indicate that the the latency can be reduced by 68%(30s -> 9.5s) to get metadata of 92 partitions.
Every partition stores 1000 files.
before:
![image](https://user-images.githubusercontent.com/4392280/134798562-fe837030-33a3-4145-9834-10895aa0c0c0.png)
after:
![image](https://user-images.githubusercontent.com/4392280/134798565-f19f6522-088b-4dca-8cab-04ca29732bf0.png)

Because we get metadata in a cache. we take the time of the first query minus the time of the second query as the metadata query time.
therefore it is about 30s -> 9.5s about latency reduced by 68% for getting metadata.